typed action

+++ src/App.tsx
@@ -40,9 +40,13 @@ function App() {
     }
   }, [dispatch, comments]);
 
-  const removeReplyHandler = () => dispatch(replyRemoved(comments, deleteId));
-  const removeCommentHandler = () =>
-    dispatch(commentRemoved(comments, deleteId));
+  const removeReplyHandler = () => {
+    if (deleteId) dispatch(replyRemoved(comments, deleteId));
+  };
+  const removeCommentHandler = () => {
+    if (deleteId) dispatch(commentRemoved(comments, deleteId));
+  };
+
   // resetStore();
   ///////////////////////////////////////////////
   /////JSX
